Add script to build conda packages for 3rd party packages
And a config file listing packages we want to build

Spinapi should probably be removed from this once we make it build conda
packages itself. I'm concerned conda is modifying the shared libraries
incorrectly - we've told it it's a noarch package but I can still see
it's added tens of kBs to them. We should move to building
platform-specific conda packages for spinapi. But this will help for
now.