Attempting to build windows-curses.
This is a very difficult to compile package, so I've added a feature to
setuptools-conda that will use the setuptools project only for metadata
to make the recipe, and will use a wheel downloaded from PyPI as the
input to conda-build.

Since there is no sdist availalble for the project, we have pip create
one from github. This will break if the version on github bumps without
releasing new wheels, but we will have to deal with it then. The github
repo doesn't have tags, so we can't do something like checking out the
most recent tag. So hopefully the developer only bumps the version when
they do a release. Releases are also infrequent as the underlying
project being wrapped is mature.