Disable uploading to test label.
It seem anaconda cloud doesn't let you have the same file in both test
and main labels. So we will just have to not use the test label for the
3rd party conda builds.

We can keep doing it for our own packages since it's only dev version
numbers that are uploaded to the test channel, so they will have unique
version numbers and filenames.