Use `runner.temp` as the build directory
Although using `conda-incubator/setup-miniconda` solved the issue of
failed Windows builds when I was testing in a separate repository, it
did not solve the issue here.

The problem seems to have to do with filepath lengths. When conda-build
is setting up the build environment (setting environment variables
before entering the developer command prompt), we sometimes see errors
like:
```
The input line is too long.
The syntax of the command is incorrect.
```

The `runner.temp` variable on GitHub actions is a nice, short path that
we can use as our build directory. So instead of
`D:\a\vendored-conda-builds\vendored-conda-builds\build\windows-curses\`
as a directory of one of the projects we're building, we now have
`D:\a\_temp\windows-curses\`, much shorter.

Problem seems to be fixed for now.